test: add 46 tests for all P1/P2/P3 fixes + fix weighted RidgeCV GPU
New tests cover every fix from 3 rounds of code review:

- P1 structural: merge conflict, INTERCEPT_CLIP_BOUND, singleton, typos

- P1 solver: gamma gradient, dg init, alpha_max, n_iter, n_samples_vec

- P2 backend: sigmoid, eigval, cuda:1, cholesky

- P2 IRLS: beta_mom, eta_raw, mu clip, Armijo, inv_gauss

- P2 CV: score guard, cv_splits, inference_method, 1/C, C_max

- P2 FE: predict, R2, t-distribution, HC1, grand mean, vectorized

- P3: default_rng, best_score_, cv validation, kwargs, ddof, n_cont

Fix: weighted RidgeCV GPU path keeps y_mean as backend array (not float)

Tested: 349 passed, 0 failed on Tesla P100

+++ statgpu/linear_model/_ridge_cv.py
@@ -637,7 +637,7 @@ def _select_ridge_alpha_cv(
                         if bool(fit_intercept):
                             w_sum = float(backend.sum(sw_train))
                             X_wmean = backend.sum(X_train * sw_col, axis=0) / w_sum
-                            y_wmean = float(backend.sum(y_train * sw_train)) / w_sum
+                            y_wmean = backend.sum(y_train * sw_train) / w_sum
                             XtX = (X_train * sw_col).T @ X_train - w_sum * backend.outer(X_wmean, X_wmean)
                             Xty = (X_train * sw_col).T @ y_train - w_sum * X_wmean * y_wmean
                             X_mean = X_wmean
